# core-banking-payments-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-baking-payments
  description: Payments Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/payments
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/payment-orders/{paymentOrderId}/correspondence/{paymentCorrespondenceId}:
    get:
      tags:
        - Payment Correspondence
      summary: Get Correspondence by ID
      description: Retrieves a specific correspondence by its ID.
      operationId: getCorrespondenceById
      parameters:
        - name: paymentOrderId
          in: path
          description: Payment Order ID
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentCorrespondenceId
          in: path
          description: Payment Correspondence ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved correspondence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCorrespondenceDTO'
        '404':
          description: Correspondence not found
    put:
      tags:
        - Payment Correspondence
      summary: Update Correspondence
      description: Updates a correspondence for a payment order.
      operationId: updateCorrespondence
      parameters:
        - name: paymentOrderId
          in: path
          description: Payment Order ID
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentCorrespondenceId
          in: path
          description: Payment Correspondence ID
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCorrespondenceDTO'
        required: true
      responses:
        '200':
          description: Successfully updated correspondence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCorrespondenceDTO'
        '404':
          description: Correspondence not found
    delete:
      tags:
        - Payment Correspondence
      summary: Delete Correspondence
      description: Deletes a correspondence for a payment order.
      operationId: deleteCorrespondence
      parameters:
        - name: paymentOrderId
          in: path
          description: Payment Order ID
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentCorrespondenceId
          in: path
          description: Payment Correspondence ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted correspondence
        '404':
          description: Correspondence not found
...
  /api/v1/payment-orders/{paymentOrderId}/correspondence:
    get:
      tags:
        - Payment Correspondence
      summary: Retrieve Correspondence
      description: Retrieves a paginated list of correspondence for a specific payment order.
      operationId: getCorrespondenceByPaymentOrderId
      parameters:
        - name: paymentOrderId
          in: path
          description: Payment Order ID
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.paymentOrderId
          in: query
          description: Exact filter for payment order id
          required: false
          schema:
            type: string
        - name: filters.correspondenceType
          in: query
          description: Exact filter for correspondence type
          required: false
          schema:
            type: string
        - name: filters.correspondenceDate
          in: query
          description: Exact filter for correspondence date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[correspondenceDate].from
          in: query
          description: Filter correspondence date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[correspondenceDate].to
          in: query
          description: Filter correspondence date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.correspondenceChannel
          in: query
          description: Exact filter for correspondence channel
          required: false
          schema:
            type: string
        - name: filters.correspondenceDirection
          in: query
          description: Exact filter for correspondence direction
          required: false
          schema:
            type: string
        - name: filters.correspondentBank
          in: query
          description: Exact filter for correspondent bank
          required: false
          schema:
            type: string
        - name: filters.correspondentReference
          in: query
          description: Exact filter for correspondent reference
          required: false
          schema:
            type: string
        - name: filters.messageContent
          in: query
          description: Exact filter for message content
          required: false
          schema:
            type: string
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved correspondence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No correspondence found
    post:
      tags:
        - Payment Correspondence
      summary: Create Correspondence
      description: Creates a new correspondence for a payment order.
      operationId: createCorrespondence
      parameters:
        - name: paymentOrderId
          in: path
          description: Payment Order ID
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCorrespondenceDTO'
        required: true
      responses:
        '201':
          description: Successfully created correspondence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCorrespondenceDTO'
        '400':
          description: Invalid input
...
  /api/v1/payment-orders/{paymentOrderId}/correspondence/type/{correspondenceType}:
    get:
      tags:
        - Payment Correspondence
      summary: Retrieve Correspondence by Type
      description: Retrieves a paginated list of correspondence for a specific payment order and correspondence type.
      operationId: getCorrespondenceByPaymentOrderIdAndType
      parameters:
        - name: paymentOrderId
          in: path
          description: Payment Order ID
          required: true
          schema:
            type: string
            format: uuid
        - name: correspondenceType
          in: path
          description: Correspondence Type
          required: true
          schema:
            type: string
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.paymentOrderId
          in: query
          description: Exact filter for payment order id
          required: false
          schema:
            type: string
        - name: filters.correspondenceType
          in: query
          description: Exact filter for correspondence type
          required: false
          schema:
            type: string
        - name: filters.correspondenceDate
          in: query
          description: Exact filter for correspondence date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[correspondenceDate].from
          in: query
          description: Filter correspondence date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[correspondenceDate].to
          in: query
          description: Filter correspondence date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.correspondenceChannel
          in: query
          description: Exact filter for correspondence channel
          required: false
          schema:
            type: string
        - name: filters.correspondenceDirection
          in: query
          description: Exact filter for correspondence direction
          required: false
          schema:
            type: string
        - name: filters.correspondentBank
          in: query
          description: Exact filter for correspondent bank
          required: false
          schema:
            type: string
        - name: filters.correspondentReference
          in: query
          description: Exact filter for correspondent reference
          required: false
          schema:
            type: string
        - name: filters.messageContent
          in: query
          description: Exact filter for message content
          required: false
          schema:
            type: string
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved correspondence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No correspondence found
...
components:
  schemas:
...
    PaymentCorrespondenceDTO:
      required:
        - correspondenceChannel
        - correspondenceDate
        - correspondenceDirection
        - correspondenceType
        - messageContent
        - paymentOrderId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        paymentCorrespondenceId:
          type: string
          format: uuid
          readOnly: true
        paymentOrderId:
          type: string
          format: uuid
        correspondenceType:
          maxLength: 50
          minLength: 0
          type: string
        correspondenceDate:
          type: string
          format: date-time
        correspondenceChannel:
          maxLength: 50
          minLength: 0
          type: string
        correspondenceDirection:
          pattern: ^(INBOUND|OUTBOUND)$
          type: string
        correspondentBank:
          maxLength: 255
          minLength: 0
          type: string
        correspondentReference:
          maxLength: 100
          minLength: 0
          type: string
        messageContent:
          maxLength: 2000
          minLength: 0
          type: string
        attachmentId:
          type: string
          format: uuid
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...